+++ neural_lam/train_model.py
@@ -271,6 +271,19 @@ def main(input_args=None):
         help="Number of future time steps to use as input for forcing data",
     )
     parser.add_argument(
+        "--num_past_boundary_steps",
+        type=int,
+        default=1,
+        help="Number of past time steps to use as input for boundary forcing",
+    )
+    parser.add_argument(
+        "--num_future_boundary_steps",
+        type=int,
+        default=1,
+        help="Number of future time steps to use as input for boundary "
+        "forcing",
+    )
+    parser.add_argument(
         "--load_single_member",
         action="store_true",
         help=(
@@ -314,8 +327,10 @@ def main(input_args=None):
     # Set seed
     seed.seed_everything(args.seed)
 
-    # Load neural-lam configuration and datastore to use
-    config, datastore = load_config_and_datastore(config_path=args.config_path)
+    # Load neural-lam configuration and datastores to use
+    config, datastore, datastore_boundary = load_config_and_datastore(
+        config_path=args.config_path
+    )
 
     # Create datamodule
     data_module = WeatherDataModule(
@@ -325,6 +340,9 @@ def main(input_args=None):
         standardize=True,
         num_past_forcing_steps=args.num_past_forcing_steps,
         num_future_forcing_steps=args.num_future_forcing_steps,
+        num_past_boundary_steps=args.num_past_boundary_steps,
+        num_future_boundary_steps=args.num_future_boundary_steps,
+        datastore_boundary=datastore_boundary,
         load_single_member=args.load_single_member,
         batch_size=args.batch_size,
         num_workers=args.num_workers,